Large set of updates associated with latest paper
This is an update to nearly all facets of the Tn-seq analysis pipeline.
It includes speed improvements in the binning of sites into genes,
filtering for reads that include both the amplification primer and the
Tn end sequence, and the definition of the differential abundance
analysis as an R function, so it can be loaded into RStudio or other
IDEs for interactive analyses, figure generation, etc. It also includes
the new essential genes analysis, to be described in a new paper
shortly.