Error in the handling of duplicate reads
Previously, reads were filtered on MAPQ. However, MAPQ heavily penalizes reads for not mapping uniquely, which occurs often in repeated regions of bacterial genomes. To correct this, post-adapter trimming length was increased to 35 to increase the ability to accurately uniquely map trimmed reads, and bowtie2 was put into "-a" mode, which reports all matches. As these are reported in descending order of alignment score, with ties randomly ordered, we can simply choose the first reported match for any single read to randomly assign ambiguous reads to one of its possible origins. Thus, transposon insertions that would yield identical reads from more than one location are impossible to place correctly in the genome and should be viewed with suspicion.